Refactoring to new JSDebugger subsystem
