+++ Source/AtomicApp/Player/IPCPlayerApp.h
@@ -47,8 +47,6 @@ namespace Atomic
 
         virtual void ProcessArguments();
 
-        void Reconnect();
-
     protected:
 
     private:
